Move interceptor and authorization wrappers
