+++ ledger/ledger-api-auth/BUILD.bazel
@@ -17,18 +17,22 @@ da_scala_library(
     ],
     runtime_deps = [],
     deps = [
-        "//daml-lf/archive:daml_lf_dev_archive_java_proto",
         "//daml-lf/data",
-        "//daml-lf/transaction",
-        "//daml-lf/transaction/src/main/protobuf:transaction_java_proto",
-        "//daml-lf/transaction/src/main/protobuf:value_java_proto",
+        "//ledger-api/grpc-definitions:ledger-api-scalapb",
+        "//ledger-api/rs-grpc-akka",
+        "//ledger-api/rs-grpc-bridge",
+        "//ledger/ledger-api-akka",
+        "//ledger/ledger-api-client",
+        "//ledger/ledger-api-common",
         "//ledger/ledger-api-domain",
-        "@maven//:com_google_guava_guava",
-        "@maven//:com_google_protobuf_protobuf_java",
+        "//ledger/ledger-api-scala-logging",
         "@maven//:com_typesafe_akka_akka_actor_2_12",
         "@maven//:com_typesafe_akka_akka_stream_2_12",
         "@maven//:io_grpc_grpc_api",
+        "@maven//:io_grpc_grpc_context",
         "@maven//:io_grpc_grpc_core",
         "@maven//:io_grpc_grpc_services",
+        "@maven//:org_scala_lang_modules_scala_java8_compat_2_12",
+        "@maven//:org_slf4j_slf4j_api",
     ],
 )
