+++ ledger/ledger-api-auth/src/main/scala/com/digitalasset/ledger/api/auth/services/PackageManagementServiceAuthorization.scala
@@ -1,15 +1,14 @@
 // Copyright (c) 2019 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.platform.server.api.authorization.services
+package com.digitalasset.ledger.api.auth.services
 
-import com.digitalasset.ledger.api.auth.AuthService
 import com.digitalasset.grpc.adapter.utils.DirectExecutionContext
+import com.digitalasset.ledger.api.auth.AuthService
 import com.digitalasset.ledger.api.v1.admin.package_management_service.PackageManagementServiceGrpc.PackageManagementService
 import com.digitalasset.ledger.api.v1.admin.package_management_service._
 import com.digitalasset.platform.api.grpc.GrpcApiService
 import com.digitalasset.platform.server.api.ProxyCloseable
-import com.digitalasset.platform.server.api.authorization.ApiServiceAuthorization
 import io.grpc.ServerServiceDefinition
 import org.slf4j.{Logger, LoggerFactory}
