+++ ledger/ledger-api-common/BUILD.bazel
@@ -30,7 +30,6 @@ da_scala_library(
         "//ledger-api/rs-grpc-akka",
         "//ledger-api/rs-grpc-bridge",
         "//ledger/ledger-api-akka",
-        "//ledger/ledger-api-auth",
         "//ledger/ledger-api-client",
         "//ledger/ledger-api-domain",
         "//ledger/ledger-api-scala-logging",
