+++ ledger/ledger-api-integration-tests/src/test/itsuite/scala/com/digitalasset/platform/tests/integration/ledger/api/AuthorizationIT.scala
@@ -9,7 +9,14 @@ import java.util.UUID
 
 import com.digitalasset.daml.bazeltools.BazelRunfiles.rlocation
 import com.digitalasset.daml.lf.data.Ref
-import com.digitalasset.ledger.api.auth.{Claim, ClaimActAsParty, ClaimAdmin, ClaimPublic, Claims}
+import com.digitalasset.ledger.api.auth.{
+  AuthServiceStatic,
+  Claim,
+  ClaimActAsParty,
+  ClaimAdmin,
+  ClaimPublic,
+  Claims
+}
 import com.digitalasset.ledger.api.domain.LedgerId
 import com.digitalasset.ledger.api.testing.utils.{
   AkkaBeforeAndAfterAll,
@@ -55,7 +62,6 @@ import com.digitalasset.platform.apitesting.{
   TestCommands,
   TestIdsGenerator
 }
-import com.digitalasset.platform.server.api.authorization.auth.AuthServiceStatic
 import com.google.protobuf.ByteString
 import io.grpc.{Status, StatusException, StatusRuntimeException}
 import io.grpc.stub.StreamObserver
